feat. STEP05_시나리오 분석 및 작업 계획, STEP06_시나리오 설계 자료

+++ src/main/java/com/concert_reservation/api/business/model/dto/command/PointCommand.java
@@ -11,7 +11,7 @@ import lombok.*;
 @AllArgsConstructor
 public class PointCommand {
     private String userId;
-    private BigDecimal amount;
+    private Long amount;
 
     public Point toEntity(User user) {
         return Point.builder()
